refactor + tweak checkbox behaviours
